+++ README.md
@@ -33,7 +33,8 @@
 <img width="791" alt="스크린샷 2022-10-01 14 49 04" src="https://user-images.githubusercontent.com/56791347/193398670-63b6989c-0d2d-45b7-ad82-d9884afacf13.png">
 
 🤍**간단한 회고**🤍  
-음...모델링하는거 생각보다 어렵다.. 학부생때도 모델링하는거에 익숙치 않았음 + 다 까먹음의 콜라보인데, 막상 다시 하려니까 제대로 한 것인지도 모르겠고 바보가 된 것 같은 느낌이다!
+1. 음...모델링하는거 생각보다 어렵다.. 학부생때도 모델링하는거에 익숙치 않았음 + 다 까먹음의 콜라보인데, 막상 다시 하려니까 제대로 한 것인지도 모르겠고 바보가 된 것 같은 느낌이다!
 갑자기 난이도가 확 올라간 느낌이라 좀 많이 당황해서.. 계속 내가 제대로 한 것인지 모르겠다라는 생각뿐이다..
-사실 학부생때는 백엔드 공부는 깨짝깨짝 해보고, 네트워크 및 인프라 공부만 했어서 이런 모델링 부분은 처음이지만 다른 백엔드 파트분들의 모델링을 보면서 공부해봐야할 것 같다..
-아직 나는 말하는 감자인듯..하다..🥔
+사실 학부생때는 백엔드 공부는 깨짝깨짝 해보고, 네트워크 및 인프라 공부만 했어서 이런 모델링 부분은 처음이지만 다른 백엔드 파트분들의 모델링을 보면서 공부해봐야할 것 같다.
+
+3. 아직 나는 말하는 감자인듯..하다..🥔 열심히 살아야겠다...
